\docType{class}
\name{Idol}
\alias{Idol}
\title{Idol class}
\description{
  Idol Class (description)
}
\details{
  \code{Idol} class is a reference class (details).
}
\section{Fields}{
  \itemize{ \item first.name. Her first name \item
  last.name. Her last name \item height. Her body height in
  cm \item weight. Her body weight in kg \item bust. Her
  top bust size in cm \item waist. Her waist size in cm
  \item hip. Her hip size in cm }
}

\section{Contains}{
  NULL
}

\section{Methods}{
  \itemize{ \item effective.capsize \item
  substantial.aspect }
}
\keyword{documentation}

