\docType{data}
\name{idolmaster765}
\alias{idolmaster765}
\title{Idols in 765 Production}
\format{A data object with three idol objects}
\description{
  A dataset containing idols with her official profile. A
  data has three typical idols in 765 Production (Hibiki,
  Azusa and Chihaya). The variables are as follows:
}
\details{
  \itemize{ \item first.name. Her first name \item
  last.name. Her last name \item height. Her body height in
  cm \item weight. Her body weight in kg \item bust. Her
  top bust size in cm \item waist. Her waist size in cm
  \item hip. Her hip size in cm }
}
\references{
  \url{http://www.bandainamcogames.co.jp/cs/list/idolmaster/}
}
\keyword{datasets}

