% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDOLoptimize.R
\name{IDOLoptimize}
\alias{IDOLoptimize}
\title{IDOLoptimize}
\usage{
IDOLoptimize(candDMRFinderObject, trainingBetas, trainingCovariates,
  libSize = 300, maxIt = 500, numCores = 4)
}
\arguments{
\item{candDMRFinderObject:}{List object returned from the CandidateDMRFinder.v2 function}

\item{trainingBetas:}{A J x N matrix of beta-values where methylation was profiled 
in a heterogenous tissue type (i.e., WB).  Here, J indicates
the number of CpGs and N, the number of samples}

\item{trainingCovariates:}{A N x P data.frame of meta data aross the N samples.  Contained
within the meta data must be the observed cell fractions (i.e. FACS or
otherwise) for the K cell types indicated in the list object returned 
from the CandidateDMRFinder.v2 function.  The naming of cell types
should be consistent between these two objects as well.}

\item{libSize:}{Size of the optimized IDOL library.  Defaults to 300.}

\item{maxIt:}{Maximum number of iterations for the IDOL algorithm. Defaults 
to 500.}

\item{numCores:}{Number of processing cores to use (see R package DoParallel).
Defaults to 4.}
}
\value{
A list containing six objects: (1) "IDOL Optimized Library" - a vector containing the names 
           of the CpGs in the identified IDOL optimized library (2) "IDOL Optimized CoefEsts" - 
           matrix of the within-cell type mean methylation beta values across the CpGs in the
           optimal library (3) "RMSE" average root-mean squared error calculated across each 
           iteration of IDOL (4) "R2" average R2 (coefficent of determinatino) calculated across each 
           iteration of IDOL (5) "Number of iterations" how many iterations of IDOL were used
           (6) "Library Size" libsize above.
}
\description{
This function identifies optimal DML/DMR libraries for cell mixture deconvolution using 
   the procedure described in Koestler et al., (2016).
}
