% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectWBCnew.R
\name{projectWBCnew}
\alias{projectWBCnew}
\title{projectWBCnew}
\usage{
projectWBCnew(Y, coefWBC, contrastWBC = NULL, nonnegative = TRUE)
}
\arguments{
\item{Y:}{A J x N matrix of methylation beta-values collected from mixed/
heterogeneous biospecimen (i.e., WB).  Target set.}

\item{coefWBC:}{A J x K projection matrix;, i.e., within-cell type mean methylation 
matrix across J DMLs and K many cell types}

\item{contrastWBC:}{Contrast for cell composition predictions.  The user needn't modify
                   this 

nonnegative:     Should cell predictions be nonnegative.  Defaults to TRUE}
}
\value{
A N x K matrix of cell proportion estimates across the K cell types for each 
           of the N subjects contained in the Target Set.
}
\description{
This function predicts the underlying cellular composition of heterogeneous tissue 
   types (i.e., WB) using the constrained projection procedure described by Houseman et al., 
   (2012).
}
