\name{effective.capsize}
\alias{effective.capsize}
\title{Calcuate effective bust cap size ratio}
\arguments{
  \item{object}{idol object}
}
\description{
  This function calcuate effective bust cap size ratio of
  your idol when you can know her cap size.
}
\examples{
my.idol     <- new("Idol")
my.idol.ecr <- effective.capsize(my.idol)
}
\keyword{bust}

