\name{substantial.aspect}
\alias{substantial.aspect}
\title{Calcuate substantial aspect ratio}
\arguments{
  \item{object}{idol object}
}
\description{
  This function calcuate substantial aspect ratio of your
  idol This statistic represents her body shape.
}
\examples{
my.idol     <- new("Idol")
my.idol.sbr <- substantial.aspect(my.idol)
}
\keyword{body}
\keyword{shape}

